# backend/openapi.json
{
  "components" : {
    "schemas" : {
      "AttendeeDto" : {
        "properties" : {
          "avatarUrl" : {
            "type" : "string"
          },
          "instant" : {
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "ownerSnowflake" : {
            "type" : "string"
          },
          "snowflake" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AttendeeStatDto" : {
        "properties" : {
          "avatarUrl" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "eventCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "EventCategoryDto" : {
        "properties" : {
          "eventCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "totalAttendees" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "EventSummaryDto" : {
        "properties" : {
          "dateTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GalleryAlbumDto" : {
        "properties" : {
          "albumId" : {
            "type" : "string"
          },
          "albumUrl" : {
            "type" : "string"
          },
          "assetCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "attendees" : {
            "items" : {
              "$ref" : "#/components/schemas/AttendeeDto"
            },
            "type" : "array"
          },
          "eventDateTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "eventId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "eventName" : {
            "type" : "string"
          },
          "thumbnailUrl" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "GraphEdgeDto" : {
        "properties" : {
          "sharedEvents" : {
            "format" : "int32",
            "type" : "integer"
          },
          "user1Snowflake" : {
            "type" : "string"
          },
          "user2Snowflake" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "GraphNodeDto" : {
        "properties" : {
          "avatarUrl" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "eventCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "snowflake" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RewindStatsDto" : {
        "properties" : {
          "averageGroupSize" : {
            "format" : "double",
            "type" : "number"
          },
          "embeddingsAvailable" : {
            "type" : "boolean"
          },
          "eventsByDayOfWeek" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "object"
          },
          "eventsByMonth" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "object"
          },
          "firstEvent" : {
            "$ref" : "#/components/schemas/EventSummaryDto"
          },
          "lastEvent" : {
            "$ref" : "#/components/schemas/EventSummaryDto"
          },
          "socialGraph" : {
            "$ref" : "#/components/schemas/SocialGraphDto"
          },
          "topAttendees" : {
            "items" : {
              "$ref" : "#/components/schemas/AttendeeStatDto"
            },
            "type" : "array"
          },
          "topCategories" : {
            "items" : {
              "$ref" : "#/components/schemas/EventCategoryDto"
            },
            "type" : "array"
          },
          "topOrganizers" : {
            "items" : {
              "$ref" : "#/components/schemas/AttendeeStatDto"
            },
            "type" : "array"
          },
          "totalEvents" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalPlusOneGuests" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalRsvps" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalUniqueAttendees" : {
            "format" : "int32",
            "type" : "integer"
          },
          "year" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "SocialGraphDto" : {
        "properties" : {
          "edges" : {
            "items" : {
              "$ref" : "#/components/schemas/GraphEdgeDto"
            },
            "type" : "array"
          },
          "nodes" : {
            "items" : {
              "$ref" : "#/components/schemas/GraphNodeDto"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "StreamingResponseBody" : { },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/gallery" : {
      "get" : {
        "description" : "Returns albums for events the user attended",
        "operationId" : "getGallery",
        "parameters" : [ {
          "in" : "query",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GalleryAlbumDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "List gallery albums",
        "tags" : [ "Gallery" ]
      }
    },
    "/gallery/thumbnail/{albumId}" : {
      "get" : {
        "description" : "Streams the album cover thumbnail from Immich on demand — never buffered to disk or DB",
        "operationId" : "getAlbumThumbnail",
        "parameters" : [ {
          "in" : "path",
          "name" : "albumId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamingResponseBody"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Get album thumbnail",
        "tags" : [ "Gallery" ]
      }
    },
    "/gallery/{albumId}/open" : {
      "get" : {
        "description" : "Generates a per-user 1-week Immich share link and returns its public URL as JSON",
        "operationId" : "openAlbum",
        "parameters" : [ {
          "in" : "path",
          "name" : "albumId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Resolve an album's share URL for the current user",
        "tags" : [ "Gallery" ]
      }
    },
...
    "/rewind" : {
      "get" : {
        "operationId" : "getGuildStats",
        "parameters" : [ {
          "in" : "query",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "year",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RewindStatsDto"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "tags" : [ "Rewind" ]
      }
    },
    "/rewind/me" : {
      "get" : {
        "operationId" : "getMyStats",
        "parameters" : [ {
          "in" : "query",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "year",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RewindStatsDto"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "tags" : [ "Rewind" ]
      }
    },
    "/rewind/years" : {
      "get" : {
        "operationId" : "getYears",
        "parameters" : [ {
          "in" : "query",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "format" : "int32",
                    "type" : "integer"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "tags" : [ "Rewind" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}